# 4.x: Improve the performance of Range()