@akarnokd 2018-06-28
Improvements:

 N | Time | Memory
---|------|-----------
1 | 29% | 39% 
10 | 40% | 46%
100 | 48% | 50%
1000 | 50% | 50%
10000 | 50% | 50%
100000 | 48% | 50%
1000000 | 37% | 50%